requirements.txt
pyyaml==6.0
flake8==5.0.4